requirements.txt
Pillow==6.2.1
pycparser==2.19
pyenchant==2.0.0
pytesseract==0.3.0
tesseract==0.1.3

